// src/lib/interfaces/notification-type.enum.ts
export enum NotificationType {
  SUBSCRIPTION_APPLICATION_STATUS_REVIEWING = 'SUBSCRIPTION_APPLICATION_STATUS_REVIEWING',
  SUBSCRIPTION_APPLICATION_STATUS_TO_AMMEND = 'SUBSCRIPTION_APPLICATION_STATUS_TO_AMMEND',
  SUBSCRIPTION_APPLICATION_STATUS_REJECTED = 'SUBSCRIPTION_APPLICATION_STATUS_REJECTED',
  SUBSCRIPTION_APPLICATION_STATUS_PAYMENT_PENDING = 'SUBSCRIPTION_APPLICATION_PAYMENT_PENDING',
  SUBSCRIPTION_APPLICATION_STATUS_PAYMENT_CANCELED = 'SUBSCRIPTION_APPLICATION_PAYMENT_CANCELED',
  SUBSCRIPTION_APPLICATION_STATUS_PAYMENT_CONFIRMED = 'SUBSCRIPTION_APPLICATION_PAYMENT_CONFIRMED',
  SUBSCRIPTION_APPLICATION_STATUS_PENDING = 'SUBSCRIPTION_APPLICATION_STATUS_PENDING',
}